В этой ревизии внесены следующие изменения: 1) Теперь при импорте файлов вместо культурных настроек среды по умолчанию используются инвариантные 2) Теперь список кодов оплат берется из справочников, а не из предыдущих параметров расчета 3) Теперь после окочательного утверждения параметров расчета происходит обновление справочника кодов оплат 4) Теперь алгоритм приведения матриц учета окончательно переведен на System.Decimal 5) Теперь в справочник кодов оплат добавлено поле "Тип", иллюстрирующее отношение кода оплаты к резерву или командировке 6) Теперь в справочнике коды за номером "104" и "147" назначены командировочными по умолчанию, при этом все остальные назначены резервными 7) Исправлено множество мелких неполадок и недоработок, оптимизированы некоторые циклы, логики и т.п.

+++ trunk/src/NpoMash.Erm/NpoMash.Erm.Hrm/NpoMash.Erm.Hrm.Tests/NpoMash.Erm.Hrm.Tests/StructuralTests/BringingLogicTests/OptionsBringingLogic.cs
@@ -40,8 +40,8 @@ namespace NpoMash.Erm.Hrm.Tests.StructuralTests.BringingLogicTests {
             var random = new Random();
             HrmTimeSheetLogic.TaskSheetInit(object_space, task);
             foreach (var column in object_space.GetObjects<HrmMatrixColumn>()) {
-                Int64 dep_sum = 0;
-                Int64 controlled_dep_sum = 0;
+                Decimal dep_sum = 0;
+                Decimal controlled_dep_sum = 0;
                 foreach (var cell in column.Cells) {
                     dep_sum += cell.Time;
                     if (cell.Row.Order.TypeControl == FmCOrderTypeControl.TRUDEMK_FOT) {
